// src/services/admin/patient/interfaces/request.type.ts
export interface ICreatePatientPayload {
    name: string;
    username: string;
    birth_place: string;
    birth_date: string;
    nik: string;
    email: string;
    phone_number: string;
    religion: string;
    gender: string;
    marital_status: string;
    education: string;
    job: string;
    province_id: string;
    sub_district_id: string;
    city_id: string;
    village_id: string;
    additional_address: string;
}